Added some about me info and updated the page to look better
HTML:

Updated the three sections to each be a different topic about me. Added a div to each to act as a container for the content. Also added an Image in the "Who am I" section.

CSS:

Gave each of the column headers a gradient background to make it look better.